extract captureThumbnailFromSvg into utils/thumbnail.js
