+++ apps/src/util/thumbnail.js
@@ -3,12 +3,58 @@
  * of code studio apps.
  */
 
+import {canvasFromImage, canvasToBlob, imageFromURI, svgToDataURI} from '../imageUtils';
+import {getStore} from '../redux';
+import project from '../code-studio/initApp/project';
+
 // Export html2canvas as a non-default so that it can be stubbed by tests.
 import html2canvas from 'html2canvas';
 export {html2canvas};
 
 // Thumbnail image width and height in pixels.
 const THUMBNAIL_SIZE = 180;
+
+/**
+ * @type {number} The last time at which a screenshot capture was attempted.
+ */
+let lastCaptureTimeMs = 0;
+
+/**
+ * Converts the contents of an SVG element into an image, shrinks it to
+ * width and height equal to THUMBNAIL_SIZE, and saves it to
+ * @param {SVGElement} svg Svg element to capture the contents of.
+ * @param {number} minCaptureIntervalMs Minimum allowable time between
+ *   thumbnail image captures in milliseconds.
+ */
+export function captureThumbnailFromSvg(svg, minCaptureIntervalMs) {
+  const {isShareView, isEmbedView} = getStore().getState().pageConstants;
+  if (!project.getCurrentId() || !project.isOwner || isShareView || isEmbedView) {
+    return;
+  }
+
+  // Skip capturing a screenshot if we just captured one recently.
+  if (Date.now() - lastCaptureTimeMs < minCaptureIntervalMs) {
+    return;
+  }
+  lastCaptureTimeMs = Date.now();
+
+  if (!svg) {
+    console.warn(`Thumbnail capture failed: svgStudio not found.`);
+    return;
+  }
+
+  if (!svg.toDataURL) {
+    console.warn('Thumbnail capture failed: svg.toDataURL undefined.');
+    return;
+  }
+
+  svgToDataURI(svg)
+    .then(imageFromURI)
+    .then(canvasFromImage)
+    .then(createThumbnail)
+    .then(canvasToBlob)
+    .then(project.saveThumbnail);
+}
 
 /**
  * Paint an image of an existing canvas onto a new thumbnail canvas. If the
